№3 refactor classic logic to CA
